requirements.txt
asyncio
pyrogram
pyromod
python-dotenv
hachoir
html_telegraph_poster
TgCrypto
aiohttp
